requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
cssutils==1.0.2
daphne==2.5.0
Django==3.1
django-appconf==1.0.4
django-ckeditor==5.9.0
django-filter==2.3.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-simple-mail==2.3.3
djangorestframework==3.11.1
hiredis==1.1.0
html2text==2020.1.16
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
lxml==4.5.2
Markdown==3.2.2
msgpack==1.0.0
pilkit==2.0
Pillow==7.2.0
premailer==3.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
zope.interface==5.1.0
